// src/enums/dashboard-module.enum.ts
export enum DashboardModule {
  USERS = "UsersManagement",
  CATEGORIES = "CategoriesManagement",
  SUB_CATEGORIES = "SubCategoriesManagement",
  PRODUCTS = "ProductsManagement",
  LOGOS = "LogosManagement",
  BANNERS = "BannersManagement",
  TYPE_HINT_CONFIGS = "TypeHintConfigsManagement",
  SHOWCASES = "ShowcasesManagement",
  ORDERS = "OrdersManagement",
}